+++ 10_XSD/PARS_XSD_Test/PARS_XSD_Test/parsScheme.cs
@@ -157,7 +157,7 @@ public partial class InputDataPointsPointRobot {
     
     /// <remarks/>
     [System.Xml.Serialization.XmlAttributeAttribute()]
-    public byte Scale;
+    public decimal Scale;
 }
 
 /// <remarks/>
